made all floats doubles (floats suck) also refactored drive
so we now have some baseline code in drive that inits the mechanum subsystem and then runs it

+++ TeamCode/src/main/java/org/firstinspires/ftc/team7316/util/Constants.java
@@ -5,21 +5,21 @@ package org.firstinspires.ftc.team7316.util;
  */
 public class Constants {
 
-    public static final float JOYSTICK_DRIVE_DEADZONE = 0.03f;
-    public static final float DRIVER_MOTOR_DEADZONE = 0.18f;
+    public static final double JOYSTICK_DRIVE_DEADZONE = 0.03;
+    public static final double DRIVER_MOTOR_DEADZONE = 0.18;
     public static final int ENCODER_TICK_PER_REV = 1120;
     public static final int DRIVE_RPM_MAX = 280; // can change later actual max 320
     public static final double DISTANCE_PER_REV = 4*Math.PI;
     public static final double COLOR_DIFFERENCE = 2;
 
     //Drive Base PID Constants
-    public static final float encoderP = 0;
-    public static final float encoderI = 0;
-    public static final float encoderD = 0;
+    public static final double encoderP = 0;
+    public static final double encoderI = 0;
+    public static final double encoderD = 0;
 
-    public static final float gyroP = 0;
-    public static final float gyroI = 0;
-    public static final float gyroD = 0;
+    public static final double gyroP = 0;
+    public static final double gyroI = 0;
+    public static final double gyroD = 0;
 
     public static final double sqrt2 = Math.sqrt(2);
